Awspec::Generator::Doc -> Awspec::Generator::Doc::Type
