+++ lib/awspec/generator/doc/type/ebs.rb
@@ -1,0 +1,17 @@
+module Awspec::Generator
+  module Doc
+    module Type
+      class Ebs < Base
+        def initialize
+          super
+          @type_name = 'EBS'
+          @type = Awspec::Type::Ebs.new('my-ebs')
+          @ret = @type.volume
+          @matchers = []
+          @ignore_matchers = []
+          @describes = []
+        end
+      end
+    end
+  end
+end
