+++ lib/awspec/generator/doc/type/rds_db_parameter_group.rb
@@ -1,0 +1,16 @@
+module Awspec::Generator
+  module Doc
+    module Type
+      class RdsDbParameterGroup < Base
+        def initialize
+          super
+          @type = Awspec::Type::RdsDbParameterGroup.new('my-rds-db-parameter-group')
+          @ret = @type.parameters
+          @matchers = []
+          @ignore_matchers = []
+          @describes = []
+        end
+      end
+    end
+  end
+end
