+++ lib/awspec/generator/doc/type/route53_hosted_zone.rb
@@ -1,0 +1,16 @@
+module Awspec::Generator
+  module Doc
+    module Type
+      class Route53HostedZone < Base
+        def initialize
+          super
+          @type = Awspec::Type::Route53HostedZone.new('example.com.')
+          @ret = @type.hosted_zone
+          @matchers = []
+          @ignore_matchers = []
+          @describes = []
+        end
+      end
+    end
+  end
+end
