+++ lib/awspec/generator/doc/type/s3.rb
@@ -1,0 +1,16 @@
+module Awspec::Generator
+  module Doc
+    module Type
+      class S3 < Base
+        def initialize
+          super
+          @type = Awspec::Type::S3.new('my-bucket')
+          @ret = @type.bucket
+          @matchers = []
+          @ignore_matchers = []
+          @describes = []
+        end
+      end
+    end
+  end
+end
