+++ lib/awspec/generator/doc/type/vpc.rb
@@ -1,0 +1,17 @@
+module Awspec::Generator
+  module Doc
+    module Type
+      class Vpc < Base
+        def initialize
+          super
+          @type_name = 'VPC'
+          @type = Awspec::Type::Vpc.new('my-vpc')
+          @ret = @type.vpc
+          @matchers = []
+          @ignore_matchers = []
+          @describes = []
+        end
+      end
+    end
+  end
+end
